fix dollar price and cart page and detail_cat page

+++ database/migrations/2023_10_09_165433_products.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('product_name', 500);
             $table->string('product_nickname', 500);
             $table->string('product_image', 500)->nullable();
-            $table->integer('product_price');
+            $table->decimal('product_price',10);
             $table->integer('product_force_price')->default(0);
             $table->integer('cat_id');
             $table->integer('inventory')->default(1);
